Added kernel configuration fragments and ramdisk creation process (busybox fetch/configure/install + packing ramdisk)
This commit is intended to automate everything "manually" created.
As per the kernel configurations - we COULD provide smaller fragments starting with allnoconfig or tinyconfig that would suffice for a boot,
but then we would have to add every time configs (e.g. even trivial things like CONFIG_PRINTK, CONFIG_DEVTMPFS, CONFIG_SYSFS, ...) which a short talk such as this one in San Diego simply would not give
the attendees time to add configuration and rebuild.

In "real courses" do no not reveal this commit, but rather work with them "one by one" (same applies for booting without a ramdisk, without init, with "flawed init" etc.
There is a comment in the generated init about using hotplug. This usually serves as a udev/ueventd/mdev/sysfs enumeration exercise in some of the PSCG's kernel courses, so to make my life easier,
there may be a non-harming failure when writing to the /proc/sys/kernel/hotplug file. Don't worry about it.