Minimal kernel (VGA console + Keyboard + Ramdisk + ELF) + supporting minimal rootfs
In this version we illustrate a very simple /sbin/init (ELF) execution, and probvide an minimal file for /etc/inittab
This commit's configuration snappet does not support BINFMT_SCRIPT, in order to illustrate what makes the "shebang" (e.g. "#!/bin/sh") work.
So you can either see the running of /init fail(and observe the fallback to /sbin/init), or directly specify "rdinit=/sbin/init" to the kernel cmdline (if so, make sure you use rdinit - you execute from the ramdisk. Do not use init=, it will not have effect).
To get BINFMT_SCRIPT, just add it to the configuration (which is exactly done in the kernel's ef634bdbadfdd0f1c7671c724727f8d10d303842 )

From now on, we will specify the matching commits on the respective kernel tree.
kernel-configs/our-tiny-config.config is a symbolic link to the current relevant copy.
We will not have a file under kernel-configs for each commit. Instead we will add only meaningful such files, or create a subfolder for them (too much files, too much mess, and they are really unnecessary, as we explain very well how to reconstruct them, and what are the important phase. It all depends on the amount of time the student has to go through failurs till the great success.

Notes (w.r.t to the kernel tree illustration):

- Tested on 32 bits on sall of the "strategic" commmits in the kernel configuration repository.

- Testted on 64 bits at 989458bb3ef973043b5f4945d7df6ebdabb3591d  (which takes the "final, good enough 32bit" config, and just adds CONFIG_64BIT (and its automatic dependencies)

- Display and working initramfs Matches:
  commit b5d7fafe2bae24147ee74e6dd9976bb8ebca0d07 (tag: working-elf-init, wip-and-adding-serial)
- Keyboard addition (so you can actually also work with it) Mathces:
  commit 0db61092751ecc75653244c04873ad85b29d16f6 (HEAD -> ramdisk-and-init-work, tag: console-input-and-output-complete)
  This is kernel-configs/printk-vt-vga_console-keyboard-initramfs_gzip-elf.config
- For BINFMT_SCRIPT addition see:
  commit ef634bdbadfdd0f1c7671c724727f8d10d303842 (HEAD -> ramdisk-and-init-work, tag: init-works-for-elf-and-scripts)