Updated busybox binary and modified init to focus on working on console
The reason for this is that in thtis way it is easier to illustrate perfectly working version without mounting anything (i.e. no /proc /dev/ sys , ...), and show the differences and different fallbacks of the kernel init.

Also added the "dontbuild" option to fetch_busybox_and_createt_initial_ramdisk.sh . Of course it is better to have different scripts, but I didn't want to have too many. when this argument is provided the script work as follows:
- It does not fetch and unpack the busybox tarball.
- If busybox is already built - it justs reinstall it.
- Otherwise - it actutally does build it, which is the behavior we want - however it will only build for 64bit (FIXME)

Note: since this repository is mostly used for training and the public talks are excerpts of the training, we will add and remove features according to the talk / interest / etc.  So sometimes it may seem to you as if "we made a progress, but take a step back". Keep calm, and follow the process :-)