Using olddefconfig for the tiny-branch.
The reason is that in our building process of the examples we used 32-bit kernel. However, we assume the user
may want to build 64-bit kernels (especially since it's the default ARCH in envsetup.sh :-) ) and would not bother with changing the default values, so we just take the defaults for all.

Pay close attention to the commits. On the current commit, you are not supposed to be working with serial, and your init cannot be a shell script. So "failure" is deliberate, and of course, with a minor change you wil get it to work.

RUN WITH:
$ qemu-system-x86_64     -kernel ${KERNEL_OUT}/arch/x86/boot/bzImage -initrd ${RAMDISK_PACKED}  -append " " -enable-kvm

And troubleshoot (with the instructor, alone, whatever works for you :-) ).
Also, don't try to build modules as they are disabled in your (tiny) kernel.
The easy way will be on the respective conference branch or on master.