Added /usr/bin/psplash - a hacky psplash customization
This is an adjustment (OK, a serious contender for the quicekst and dirtiest hack in the history) of psplash from the yocto project.
To run it, add it to your rootfs (you can use the ./add_rootfs_extras.sh script).
To disable the splash screen, boot the kernel with the "nosplash" param.

For those who know the project the modifications here are:
- Instead of waiting for commands or relying on vc events, the program just exits after showing progress for some seconds (obviously it's completely made up)
- Shows how to draw some bitmaps.

You can start it as many times as you want from userspace, as it is a userspace splash screen!

The process for those interested is listed in a video series I posted on youtube that starts with https://www.youtube.com/watch?v=lqIbu2qzNPo&list=PLBaH8x4hthVysdRTOlg2_8hL6CWCnN5l-&index=5 .

I am not a youtuber, do not expect professional editing - but do expect professional content.