Added VT's in /init and running the splash screen from it in the background
We open here a shell on tty1,tty2,tty3,tty4 (with different colors), tty42
You also illustrate different background color strategies, so if you open also tty5 and tty6, you will see that they are also colored.
The color scheme is intentional - we use the corressponding 40+i background color where i is the tty number (i.e. 44 for /dev/tty4)

The splash screen is run in the background to allow you a safe way to let init progress even if you did not provide a console
It is recommended to provide also a serial console

Run (for example) with:
qemu-system-x86_64 -kernel ${KERNEL_OUT}/arch/x86/boot/bzImage -enable-kvm -initrd ${RAMDISK_PACKED}   -append " vga=0x312  nokaslr  console=tty0 console=ttyS0 "  -serial mon:stdio

NOTE: The (at least) next couple of commits, we will update the contents of wip_ramdisk so that it will be easier for you to
reproduce, as we guide you thorugh the differences in our different init schemes.
This means that you need to not forget to repack the ramdisk.

We will later remove the basic /init generated in fetch_busybox_and_create_initial_ramdisk.sh , and also add it to the rootfs-additions.
This will make merging to the master branch not ideal for conferences where we want to illustrate the basics, but this is the tradeoff between simplicity (and readability!) and features

No more disclaimers from this point.