Moved the splash screen for after opening ttys
To understand the rationale, modify the code in this commit so that the splash is not run in the background.
If there is no serial console (as in a PC) - you will not be able to exit the splash screen unless you either:
1. chvt
2. Some other process terminates it (or it runs with a timeout, or with the FAKE_LOGIC)

Otherwise, if there is a serial console (as the last cosnole= parameter), and we don't run psplash in the background, our serial console will be stuck until psplash exits. On the other hand, if want to release the frame buffer console from another tty, [you could run it with
--no-console-switch so that chvt does not terminate it, but the implementation will not redraw the background once we clear it when have laready executed a script (because of openvt -c <term-nunmber> which executes the shell, and /ronenv in our script)], chvt will exit the console. So we just run it in the background now and that's it.
A later commit will use busybox's init mechanism to execute some of these things the proper way (e.g. exactly like in yocto)

Now, in order to chvt (e.g. alt+f1/f2/... ) - you must first allocate the vt. So it must be done before the splash screen opens.

This is a good opportunity for you to also see what happens when you provide another graphic console, and also observe psplash code.

In our current implementation, we provide  ENV=/ronenv  both to the tty execution start, and to   the exec line in /init, so on the console, you will see the banner (because we print it in ronenv) twice.