Removed /etc/inittab in order to explain busybox init (will be restored and added to in the next commit!)
Detour
-------
We already saw (24872c8a6053be197ce454b163b9851cc65142cc) what happens when we want to have a very minimal inittab.
We had a shell running on the console, and we avoided many automatic error messages (as we'll see below), and kept just one annoying message "/bin/sh: can't access tty; job control turned off" that is harmless, and resolved by either running the shell with a "setsid -c /bin/sh" or running the ctty hack. The reason for it is that the console is not a controling terminal by default,
and you should not care about it too much, so it is ignored henceforth, so we can focus.

I will just say where it does affect you. In Job Control. For example, if you run vi and then expect to do ^Z (to have SIGTSTOP) sent to your proccess - it will not work for you unless your (sh) session has a Controling tty, because the controlling tty (which interacts with the user) expects to capture the signal and send it to a process that is the "Session Leader". When you start init (or busybox) from the console as is, this is not taken care of, so you cannot do things like handling job control, ^C etc.

Back to our point
------------------
Now let's revisit busybox's init, and build it accordingly, step by step.

When start busybox init ( rdinit=/sbin/init ), and do not provide the /etc/inittab file, we get a console, and repeating messages of
can't open /dev/tty3: No such file or directory
can't open /dev/tty2: No such file or directory
can't open /dev/tty4: No such file or directory

Nothing will be mounted so there will not be /proc , so you can't, e.g. do something like ps, and nothing will exist in /dev/

This aligns well with the default init actions as per https://git.busybox.net/busybox/tree/examples/inittab :

If you exit the shell (e.g. ^D or EOF) you will have the following message output to the console, due to the askfirst: clause.
Please press Enter to activate this console. /

Now the tty fields actually tell us where (in which tty) to execute whatever we want to execute, so we could use the inittab file to execute what we previously have from /init
Obviously, the controlling terminal rules from our "detour" apply.
~