Added inittab - starting simple with splash screen and no virtual consoles
In this commit we illustrate what happens when you have may have a serial and/or a graphical console, but
you do not execute them from the /init (see the command line below, we now use busybox init).
Run and observe that you have no way to provide input, and no virtual console switch.

Do know that there are uses for this (honestly there are uses for almost everything).

You can see that we do some of the things we would otherwise do in the init file. In the next commit we will fix some.

Run with:
./repack_ramdisk.sh && qemu-system-x86_64 -kernel ${KERNEL_OUT}/arch/x86/boot/bzImage -enable-kvm -initrd ${RAMDISK_PACKED}   -append "fbcon=nodefer  vga=0x312  nokaslr rdinit=/sbin/init console=tty0 console=ttyS0 quiet"  -serial mon:stdio