Added autostart of /bin/sh on ttys - now you can swithc consoles
This is an exercise though. Note the entries on /etc/inittab. Some are missing.
There are several considerations to attend when you plan what to start where.

In the same qemu running configuration we used before, psplash will go to tty2.
To understand the relations, you need to ask yourself (and modify inittab as well), what happen when there is a command running on a tty, while other command also runs there. If you know to separate you are good. If not - you may run into issues.

Also, what if you want for example to have login running?

So we'll do some of it in the next couple of commits, revert to some simple and working configuration, and by that conclude
the booting up and setting up initial ramfs part.
What do you do next? Create and customize your own distro.