Added getty usage example and resolved /tmp/ race condition
We illustrate several things in here:

We open getty on tty0 (which my change according to kernel cmd lines etc, as tty0 is the CURRENT virtual console), the login command.
You can see what it does after the splash screen is done.
The command itself aims to do several standard things, such as setting the terminal baudrate, and executing "login".
This is why, you are very likely to see either it, agetty, migetty, etc. running from system startup scripts (not just in busybox)
Since we do not have multiuser support (it is easy to provide but we need to rebuild the kernel for that), you will see that you can never get in. We will workaround that in the next commit.

The reason we put the creation of /tmp/ as once, was to give you the chance to see that pspslash does not wait for it.
Since psplash depends on the existence of /tmp/ for creating the fifo - by now you could have already enountered that some times it works and some times it doesn't. This commit makes it work deterministically as once will run only after sysinit events.