Completed the /init calls /sbin/init (busybox) flow.
Things are deliberately left not working perfectly, but these are cosmetics, for you to experience, and fix.
You can of course do whatever you want - it's your rootf now, just one caveat which is not trivial to grasp:
The current implementation of psplash opens the splash screen on tty2, so if you do respawn on the tty with ENV exported (as per /init), the banner willl display twice. An elegant (but not obvious unless you look at the change in autologinroot.sh) solution is to preserve the environment upon login.
It is not hard to fix, but I would need to explain too much about the tradeoffs in console parameters, order, vcs, etc, and writing would take too much time and space, and I do explain (and solve) in my clases.

That's it. Happy hacking. This is only the beginning - but if you have followed thoroughly everythig (and done!), you already have a solid understanding of Embedded Linux. Congratulations!