Hacking the dynamic libraries location so that a dynamically linked 64 bit "hello world" executable will work nicely.
The idea is to either:
$ objdump -p <your-exe> | grep NEEEDED   - and locate the NEEDED
libraries

or to
$ ldd <your-exe>

NOTE that if you do not have ld - you are going to have to get it - and
objdump -p does not tell you that, and also note that ldd runs the
linker so it is not a good idea to "run it on untrusted binaries"

When copying the libraries - you would want to use $ cp -L - so that the
final link is copied (e.g. what you'd get from readlink -f) - and not
the symbolic link.

This commit adds the "solution" by figuring out that hello world needs
libc - but our ramdisk does not have ld - so it is also copied.

It is imperative to understand (when there is time in class) what is
needed and what is not, and so either:
1. Understand better what linkers do (security)
2. Understand why we need build systems (embedded, tools)