Added Microsoft Visual Studio Code examples
Showing how to set up vscode for kernel debugging with our tiny-config.
This is not optimal, and the long list of includes is of course
unnecessary (as I explain in the video) - it's there for deductive
reasons.
I am unlikely to update this, but others are welcome to do so.

Video
--------
https://youtu.be/dXhm3Kdnnbg