build_kernel.sh: added an option to use a custom config, if you set a symlink to it from kernel-configs/our-tiny-config.config
Otherwise, if there is no such symlink, defconfig + the debug fragments will be used, as before (note that there may be some warnings there, as config changed since kernel 5.x, and some things, such as kvmconfig were removed. I am not bothering to change this, in case you want to work with older kernels. Remember - the purpose here is simplicity.