Updated wip_ramdisk/init for the demonstration, to both advance the splash screen, detect DRM or FB devices, detect QEMU via DMI sysfs interface, and mount a 9p filesystem if available (if not, there will be an error you can ignore on your console) called 'thepscg-mounts-you' at /mnt/host.
Do note that the script which generates it *was not* updated - deliberately. This particular init is used for demonstrations, for this branch, and in general, in the master branch we no longer keep a wip_ramdisk. After recording another talk with it in a meetup about Graphics on August 14, 2024 I think I will update the script an perhaps do some merging. Till then, be careful when you create the initial ramdisk - this particular edit is what I am looking at in the lecture