fetch_busybox_and_create_initial_ramdisk.sh: updated to use psplash only if /dev/fb0 is there, and auto accelerate it from /init. Then, exec /sbin/init (which also takes care of the cttyhack itself), which would not do anything around psplash (but the line is still commented out if you wanted it to, and changed /init correspondingly...)
Note that with that, the meetup17 demos are complete, until further notice. This means that we can (and will...) remove wip_ramdisk, as we can
build things ourselves from another repos (read: psplash), and we can
actually add the prebuilt psplash, and other "extra" files to the
ramdisk by (including creating the ramdisk):
1. ./fetch_busybox_and_create_initial_ramdisk.sh
2 ./add_rootfs_extras.sh

If these are done, there is no need to save anything in wip_ramdisk.

This has been tested with both busybox 1.31 and busyboxc 1.36.1