fetch_busybox_and_create_initial_ramdisk.sh: Updated to better represent the changes in master and ease later merge.
Note that this requires multilib (which will force you remove other distro installed cross-compilers), and i386 packages.
busybox 1.36.1 wants libcrypt which has no trivial i386 package in recent Ubuntu distros, so this is not recommended.
Better install an i386/i586/i686/... cross toolchain.
Testing note: I did not test on this commit x86/32, both because of that, and because the meetup was on the 64bit version